# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UTB Bus Ticketing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate mobile money payment (Wave, Orange Money, or MTN Money)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /payments/paydunya-webhook:
    post:
      operationId: paydunyaWebhook
      tags: [payments]
      summary: PayDunya payment notification (IPN) for bus tickets and hotel bookings. Called by PayDunya, not by the app; the state is re-checked with PayDunya's confirm API.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/PaydunyaWebhookInput"
      responses:
        "200":
          description: Notification processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Malformed notification
        "404":
          description: Unknown invoice
        "502":
          description: PayDunya could not be reached to confirm (PayDunya retries)
  /payments/paydunya-refund-webhook:
    post:
      operationId: paydunyaRefundWebhook
      tags: [payments]
      summary: PayDunya API PUSH notification (final state of a refund). Called by PayDunya, not by the app; the state is re-checked with PayDunya's check-status API.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaydunyaRefundWebhookInput"
      responses:
        "200":
          description: Notification processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Malformed notification
        "404":
          description: Unknown refund
        "502":
          description: PayDunya could not be reached to check the refund (PayDunya retries)

  # Tickets
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # Auth
...
    PaymentInput:
      type: object
      required: [tripId, passengerName, passengerPhone, paymentMethod]
      properties:
        tripId:
          type: integer
        seatId:
          type: integer
          description: Chosen seat (manual selection, adds the seat selection fee). Omit for free automatic assignment.
        passengerName:
          type: string
        passengerPhone:
          type: string
        paymentMethod:
          type: string
          enum: [wave, orange_money, mtn_money, moov_money, card]
    PaymentResponse:
      type: object
      required: [paymentId, amount, status]
      properties:
        paymentId:
          type: string
        amount:
          type: number
        status:
          type: string
        redirectUrl:
          type: ["string", "null"]
          description: PayDunya checkout page to send the customer to
        ticketId:
          type: ["integer", "null"]
        farePrice:
          type: number
        platformCommission:
          type: number
          description: Service fee added on top of the fare (shown to passengers as "Frais de service")
        seatSelectionFeePaid:
          type: number
        seatNumber:
          type: integer
    PaydunyaWebhookInput:
      type: object
      description: Sent by PayDunya; "data" holds the invoice object (status, hash, invoice.token, invoice.total_amount, custom_data…)
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: true
...
    PaydunyaRefundWebhookInput:
      type: object
      description: Sent by PayDunya (hash, status, token, withdraw_mode, amount, disburse_id, transaction_id, disburse_tx_id…)
      additionalProperties: true
...